visualization/requirements.txt
streamlit
pandas
plotly
imbalanced-learn
numpy
pillow